+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.10.0 未发布
+
+- [x] 重构: 移除 easy-peasy / redux, 使用 valtio 作为全局状态管理
+- [x] 重构: 移除 rxjs BehaviorSubject / recompose 等, 使用 valtio 全局组件
+- [x] 重构: 开启 TypeScript strictNullChecks
+
 ## v0.9.0 2022-07-02
 
 - [x] yarn -> pnpm